lightning_flash/requirements.txt
lightning-flash